move websocket types to core
