+++ packages/core/src/types/index.ts
@@ -10,3 +10,5 @@ export type FilterType<Base, Condition> = Pick<
     [Key in keyof Base]: Base[Key] extends Condition ? Key : never;
   }[keyof Base]
 >;
+
+export * from "./obs-websocket-js";
